all: add ts_omit_serve, start making tailscale serve/funnel be modular
    tailscaled tailscale combined (linux/amd64)
     29853147  17384418  31412596 omitting everything
    +  621570 +  219277 +  554256 .. add serve

Updates #17128
