+++ cmd/tailscale/cli/serve_legacy.go
@@ -1,6 +1,8 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
+//go:build !ts_omit_serve
+
 package cli
 
 import (
@@ -30,6 +32,10 @@ import (
 	"tailscale.com/util/slicesx"
 	"tailscale.com/version"
 )
+
+func init() {
+	maybeServeCmd = serveCmd
+}
 
 var serveCmd = func() *ffcli.Command {
 	se := &serveEnv{lc: &localClient}
